+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 2020-06-08
+
+### Added
+- Compatibility with ulrack/services version 3.0
+- Method to retrieve the method reflector from the object manager.
+
 ## 1.1.2 - 2020-06-03
 
 ### Fixed
@@ -30,7 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The initial implementation of the package.
 
 # Versions
-- [1.1.2 > Unreleased](https://github.com/ulrack/kernel/compare/1.1.2...HEAD)
+- [1.2.0 > Unreleased](https://github.com/ulrack/kernel/compare/1.2.0...HEAD)
+- [1.1.2 > 1.2.0](https://github.com/ulrack/kernel/compare/1.1.2...1.2.0)
 - [1.1.1 > 1.1.2](https://github.com/ulrack/kernel/compare/1.1.1...1.1.2)
 - [1.1.0 > 1.1.1](https://github.com/ulrack/kernel/compare/1.1.0...1.1.1)
 - [1.0.1 > 1.1.0](https://github.com/ulrack/kernel/compare/1.0.1...1.1.0)